Fruitless attempts at debugging the broken signal that comes out of my speakers.

+++ lib/sounds/sound.h
@@ -2,14 +2,14 @@
 
 namespace DynamicSounds{
     class SoundSource{
-        public:
-            SoundSource(unsigned int sample_rate);
+    public:
+        SoundSource(unsigned int sample_rate);
 
-            void virtual GenerateFrames(double *, int) = 0;
-            unsigned int virtual GetSampleRate() = 0;
-            void virtual SetSampleRate(unsigned int) = 0;
+        void virtual GenerateFrames(unsigned short *, int) = 0;
+        unsigned int virtual GetSampleRate() = 0;
+        void virtual SetSampleRate(unsigned int) = 0;
 
-        protected:
-            unsigned int sample_rate;
+    protected:
+        unsigned int sample_rate;
     };
 };
